08_stock_for_ec2/code/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
pkg_resources==0.0.0
pysqlite3==0.4.6
pyTelegramBotAPI==4.3.1
requests==2.27.1
urllib3==1.26.8
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
h11==0.12.0
idna==3.3
loguru==0.5.3
msedge-selenium-tools==3.141.3
numpy==1.22.0
outcome==1.1.0
pandas==1.3.5
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
win32-setctime==1.0.4
wsproto==1.0.0
